Revert "Add Apply Page"
This reverts commit e9d2b635f977853c9faea12528ff2510b99ed3ae.
